HAJ-2004

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/card/BatchMainCollectionService.java
@@ -10,7 +10,6 @@ import com.elm.shj.admin.portal.services.generic.GenericService;
 import com.elm.shj.admin.portal.services.sftp.SftpService;
 import com.jcraft.jsch.JSchException;
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -48,15 +47,19 @@ public class BatchMainCollectionService extends GenericService<JpaBatchMainColle
     private int retryPeriod;
 
 
-    @SneakyThrows
     @Async
     public void generateBatchCards(BatchCollectionVO batchCollectionVO) {
 
         // create temporary folder to save the cards before saving them in sftp server
-        String tmpdir = Files.createTempDirectory("cards").toFile().getAbsolutePath();
+        String tmpdir = null;
+        try {
+            tmpdir = Files.createTempDirectory("cards").toFile().getAbsolutePath();
+        } catch (IOException e) {
+            log.debug("failed to create temp folder");
+            return;
+        }
         Path root = Paths.get(tmpdir);
 
-        mainCollectionLoop:
         for (BatchMainCollectionDto batchMainCollectionDto : batchCollectionVO.getBatchMainCollections()) {
             Optional<JpaBatchMainCollection> batchMainCollection = batchMainCollectionRepository.findTopByReferenceNumberOrderByCreationDateDesc(batchMainCollectionDto.getReferenceNumber());
             if (!batchMainCollection.isPresent() || batchMainCollection.get().getStatusCode().equals(ECollectionStatus.FAIL_TO_GENERATE.name())) {
@@ -65,26 +68,43 @@ public class BatchMainCollectionService extends GenericService<JpaBatchMainColle
                 }
                 batchMainCollectionDto.setStatusCode(ECollectionStatus.GENERATING_CARDS.name());
                 BatchMainCollectionDto savedBatchMainCollection = save(batchMainCollectionDto);
-                log.info("batch main collection saved successfully");
+                log.info("batch main collection saved successfully {} ", batchMainCollectionDto.getReferenceNumber());
 
                 String sftpPath = "";
+                subCollectionLoop:
                 for (SubCollectionVO subCollectionVO : batchMainCollectionDto.getSubCollections()) {
 
                     for (String digitalId : subCollectionVO.getDigitalIds()) {
-                        BadgeVO badge;
-                        if (batchCollectionVO.getTarget().equals(EPrintingRequestTarget.APPLICANT.name())) {
-                            badge = badgeService.generateApplicantBadge(digitalId, false);
-                        } else {
-                            badge = badgeService.generateStaffCard(digitalId);
-                        }
-                        if (badge != null) {
-                            //TODO: surround with try catch , update status and break from outer loop
-                            byte[] decodedImage = Base64.getDecoder().decode(badge.getBadgeImage());
-                            InputStream targetStream = new ByteArrayInputStream(decodedImage);
-                            sftpPath = batchCollectionVO.getBatchReferenceNumber() + "/" + batchMainCollectionDto.getReferenceNumber() + "/" + subCollectionVO.getReferenceNumber();
-                            Path p = Files.createDirectories(root.resolve(Paths.get(sftpPath)));
-                            Files.copy(targetStream, p.resolve(Paths.get(digitalId + ".jpg")), StandardCopyOption.REPLACE_EXISTING);
-                            log.info("file uploaded successfully to: {}", sftpPath);
+                        try {
+                            BadgeVO badge;
+                            if (batchCollectionVO.getTarget().equals(EPrintingRequestTarget.APPLICANT.name())) {
+                                badge = badgeService.generateApplicantBadge(digitalId, false);
+                            } else {
+                                badge = badgeService.generateStaffCard(digitalId);
+                            }
+                            if (badge != null) {
+                                try {
+                                    byte[] decodedImage = Base64.getDecoder().decode(badge.getBadgeImage());
+                                    InputStream targetStream = new ByteArrayInputStream(decodedImage);
+                                    sftpPath = batchCollectionVO.getBatchReferenceNumber() + "/" + batchMainCollectionDto.getReferenceNumber() + "/" + subCollectionVO.getReferenceNumber();
+                                    Path p = Files.createDirectories(root.resolve(Paths.get(sftpPath)));
+                                    Files.copy(targetStream, p.resolve(Paths.get(digitalId + ".jpg")), StandardCopyOption.REPLACE_EXISTING);
+                                    log.info("file uploaded successfully to: {}", sftpPath);
+                                } catch (IOException e) {
+                                    log.debug("failed to save card in temporary folder");
+                                    savedBatchMainCollection.setStatusCode(ECollectionStatus.FAIL_TO_GENERATE.name());
+                                    save(savedBatchMainCollection);
+                                    break subCollectionLoop;
+
+                                }
+                            } else {
+                                break subCollectionLoop;
+                            }
+                        } catch (Exception e) {
+                            log.debug("failed to generate badge for uin {}", digitalId);
+                            savedBatchMainCollection.setStatusCode(ECollectionStatus.FAIL_TO_GENERATE.name());
+                            save(savedBatchMainCollection);
+                            break subCollectionLoop;
                         }
 
 
@@ -92,33 +112,41 @@ public class BatchMainCollectionService extends GenericService<JpaBatchMainColle
 
                 }
 
-                // create zip file for the main collection
-                //TODO: surround with try / catch and update status
-                Path zipSource = root.resolve(batchCollectionVO.getBatchReferenceNumber()).resolve(batchMainCollectionDto.getReferenceNumber());
-                sftpService.zipFolder(zipSource);
-                String zipPath = batchCollectionVO.getBatchReferenceNumber() + "/" + batchMainCollectionDto.getReferenceNumber() + ".zip";
-                FileInputStream fis = new FileInputStream(root + "/" + zipPath);
+                try {
+                    // create zip file for the main collection in temp folder
+                    Path zipSource = root.resolve(batchCollectionVO.getBatchReferenceNumber()).resolve(batchMainCollectionDto.getReferenceNumber());
+                    sftpService.zipFolder(zipSource);
+                    String zipPath = batchCollectionVO.getBatchReferenceNumber() + "/" + batchMainCollectionDto.getReferenceNumber() + ".zip";
+                    FileInputStream fis = new FileInputStream(root + "/" + zipPath);
 
-                for (int i = 0; i < maxRetries; i++)
-                    try {
-                        sftpService.uploadFile(zipPath, fis, CARDS_CONFIG_PROPERTIES);
-                        savedBatchMainCollection.setStatusCode(ECollectionStatus.READY.name());
-                        savedBatchMainCollection.setUrl(zipPath);
-                        save(savedBatchMainCollection);
-                        break;
-                    } catch (Exception e) {
-                        log.debug("failed to connect to sftp server , will retry ");
-                        Thread.sleep(retryPeriod);
-                        if (i == maxRetries) {
-                            log.debug("failed to connect to sftp server after achieving max retries");
-                            savedBatchMainCollection.setStatusCode(ECollectionStatus.FAIL_TO_GENERATE.name());
+                    // start auto retry
+                    for (int i = 0; i < maxRetries; i++)
+                        try {
+                            //upload main collection zip folder in sftp server and mark main collection as ready
+                            sftpService.uploadFile(zipPath, fis, CARDS_CONFIG_PROPERTIES);
+                            savedBatchMainCollection.setStatusCode(ECollectionStatus.READY.name());
+                            savedBatchMainCollection.setUrl(zipPath);
                             save(savedBatchMainCollection);
-                        }
+                            break;
+                        } catch (Exception e) {
+                            log.debug("failed to connect to sftp server , will retry ");
+                            Thread.sleep(retryPeriod);
+                            if (i == maxRetries) {
+                                log.debug("failed to connect to sftp server after achieving max retries");
+                                savedBatchMainCollection.setStatusCode(ECollectionStatus.FAIL_TO_GENERATE.name());
+                                save(savedBatchMainCollection);
+                            }
 
-                    } finally {
-                        fis.close();
-                        Files.delete(root.resolve(zipPath));
-                    }
+                        } finally {
+                            fis.close();
+                            Files.delete(root.resolve(zipPath));
+                        }
+                } catch (Exception e) {
+                    log.debug("failed to create zip folder for main collection {} ", batchMainCollectionDto.getReferenceNumber());
+                    savedBatchMainCollection.setStatusCode(ECollectionStatus.FAIL_TO_GENERATE.name());
+                    save(savedBatchMainCollection);
+                }
+
 
             }
 
@@ -128,15 +156,20 @@ public class BatchMainCollectionService extends GenericService<JpaBatchMainColle
         // create the zip file for  the whole batch
         try {
 
-            //TODO: check if files have been already created before zipping them
+            //TODO: check if files have been already created before zipping them or at least one main collection is ready
             Path zipSource = root.resolve(Paths.get(batchCollectionVO.getBatchReferenceNumber()));
-            sftpService.zipFolder(zipSource);
-            String zipPath = batchCollectionVO.getBatchReferenceNumber() + ".zip";
-            FileInputStream fis = new FileInputStream(root + "/" + zipPath);
-            sftpService.uploadFile(batchCollectionVO.getBatchReferenceNumber() + "/" + zipPath, fis, CARDS_CONFIG_PROPERTIES);
-            fis.close();
-            Files.delete(root.resolve(zipPath));
-            deleteDirectory(new File(root + "/" + batchCollectionVO.getBatchReferenceNumber()));
+            if (Files.exists(zipSource)) {
+                // create zip folder for the whole batch in temp folder
+                sftpService.zipFolder(zipSource);
+                String zipPath = batchCollectionVO.getBatchReferenceNumber() + ".zip";
+                FileInputStream fis = new FileInputStream(root + "/" + zipPath);
+                //upload zip folder in sftp server
+                sftpService.uploadFile(batchCollectionVO.getBatchReferenceNumber() + "/" + zipPath, fis, CARDS_CONFIG_PROPERTIES);
+                fis.close();
+                //delete temp folder and sub folders
+                Files.delete(root.resolve(zipPath));
+                deleteDirectory(new File(root + "/" + batchCollectionVO.getBatchReferenceNumber()));
+            }
 
 
         } catch (IOException | JSchException e) {
